Create README.md
Welcome to the Python project exploring the Evolution of Lego!

You are a Data Analyst at Lego working with the Sales/Customer Success teams. The Account Executive responsible for the Star Wars partnership has asked for specific information in preparation for their meeting with the Star Wars team. Although Star Wars was critical to the survival of the brand, Lego has since intoduced a wide variety of licensed sets over subsequent years.

Your two questions as follows:

1. What percentage of all licensed sets over released were Star Wars themed? 
   Save you answer as a variable the_force in the form of an integer (e.g.25)

2. In which year was Star Wars not the most popular licensed theme (in terms of number of sets released that year)?
   Save your answer as a variable new_era in the form of an integer (e.g.2012)